use has_*() for pkgdown reference index
add a @family tag to group functions in the
reference index for pkgdown
(https://statistikat.github.io/VIM/reference/index.html).

This is done with the has_concept() selector
in pkgdown/_pkgdown.yaml. As a side effect,
the reference pages are now cross linked in
the R documentation viewer as well as on pkgdown.

The new families are named like the groups
used in the reference index.
* plotting functions
* imputation methods

Also use the has_keyword() selector to
automatically group all datasets into a category.

has_keyword() was introduced recently in
https://github.com/r-lib/pkgdown/pull/1241.
Therefore, the development version of pkgdown is
now necessary in order to render the
pkgdown-docs for VIM.